+++ android/src/main/java/com/rubdev/zebrautil/ZebraUtilPlugin.java
@@ -4,7 +4,6 @@ import android.content.Context;
 
 import androidx.annotation.NonNull;
 
-import com.rubdev.zebrautil.helper.LocaleHelper;
 
 import java.util.ArrayList;